Organized some scripts
