+++ bci-game/Assets/Scripts/Entity.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 798aee76685e4133a893b19f81c3037d
+timeCreated: 1700437351
