+++ bci-game/Assets/Scripts/Entity/Interfaces.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: b59005f83f7745f287ef84e82a6fe2ce
+timeCreated: 1701136716
