+++ bci-game/Assets/Scripts/Entity/Player.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: de60b633741e4c98900ed4e96447200e
+timeCreated: 1700437394
